requirements.txt
asgiref==3.4.1
Django==3.2.6
django-multiselectfield==0.1.12
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.12.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
